require IChecks to receive entity context
