+++ apps/workflowengine/lib/Check/AbstractStringCheck.php
@@ -22,9 +22,9 @@
 namespace OCA\WorkflowEngine\Check;
 
 
-use OCP\Files\Storage\IStorage;
 use OCP\IL10N;
 use OCP\WorkflowEngine\ICheck;
+use OCP\WorkflowEngine\IEntity;
 use OCP\WorkflowEngine\IManager;
 
 abstract class AbstractStringCheck implements ICheck {
@@ -120,5 +120,9 @@ abstract class AbstractStringCheck implements ICheck {
 		}
 		$this->matches[$patternHash][$subjectHash] = preg_match($pattern, $subject);
 		return $this->matches[$patternHash][$subjectHash];
+	}
+
+	public function setEntitySubject(IEntity $entity, $subject): void {
+		// Noop
 	}
 }
