+++ apps/workflowengine/lib/Check/FileName.php
@@ -23,21 +23,15 @@ declare(strict_types=1);
 namespace OCA\WorkflowEngine\Check;
 
 use OCA\WorkflowEngine\Entity\File;
-use OCP\Files\Storage\IStorage;
 use OCP\IL10N;
 use OCP\IRequest;
 use OCP\WorkflowEngine\IFileCheck;
 
 class FileName extends AbstractStringCheck implements IFileCheck {
+	use TFileCheck;
 
 	/** @var IRequest */
 	protected $request;
-
-	/** @var IStorage */
-	protected $storage;
-
-	/** @var string */
-	protected $path;
 
 	/**
 	 * @param IL10N $l
@@ -46,15 +40,6 @@ class FileName extends AbstractStringCheck implements IFileCheck {
 	public function __construct(IL10N $l, IRequest $request) {
 		parent::__construct($l);
 		$this->request = $request;
-	}
-
-	/**
-	 * @param IStorage $storage
-	 * @param string $path
-	 */
-	public function setFileInfo(IStorage $storage, $path) {
-		$this->storage = $storage;
-		$this->path = $path;
 	}
 
 	/**
