+++ apps/workflowengine/lib/Check/FileSystemTags.php
@@ -25,7 +25,6 @@ namespace OCA\WorkflowEngine\Check;
 use OCA\WorkflowEngine\Entity\File;
 use OCP\Files\Cache\ICache;
 use OCP\Files\IHomeStorage;
-use OCP\Files\Storage\IStorage;
 use OCP\IL10N;
 use OCP\SystemTag\ISystemTagManager;
 use OCP\SystemTag\ISystemTagObjectMapper;
@@ -34,6 +33,7 @@ use OCP\WorkflowEngine\ICheck;
 use OCP\WorkflowEngine\IFileCheck;
 
 class FileSystemTags implements ICheck, IFileCheck {
+	use TFileCheck;
 
 	/** @var array */
 	protected $fileIds;
@@ -50,12 +50,6 @@ class FileSystemTags implements ICheck, IFileCheck {
 	/** @var ISystemTagObjectMapper */
 	protected $systemTagObjectMapper;
 
-	/** @var IStorage */
-	protected $storage;
-
-	/** @var string */
-	protected $path;
-
 	/**
 	 * @param IL10N $l
 	 * @param ISystemTagManager $systemTagManager
@@ -65,15 +59,6 @@ class FileSystemTags implements ICheck, IFileCheck {
 		$this->l = $l;
 		$this->systemTagManager = $systemTagManager;
 		$this->systemTagObjectMapper = $systemTagObjectMapper;
-	}
-
-	/**
-	 * @param IStorage $storage
-	 * @param string $path
-	 */
-	public function setFileInfo(IStorage $storage, $path) {
-		$this->storage = $storage;
-		$this->path = $path;
 	}
 
 	/**
