+++ apps/workflowengine/lib/Check/UserGroupMembership.php
@@ -22,12 +22,12 @@
 namespace OCA\WorkflowEngine\Check;
 
 
-use OCP\Files\Storage\IStorage;
 use OCP\IGroupManager;
 use OCP\IL10N;
 use OCP\IUser;
 use OCP\IUserSession;
 use OCP\WorkflowEngine\ICheck;
+use OCP\WorkflowEngine\IEntity;
 use OCP\WorkflowEngine\IManager;
 
 class UserGroupMembership implements ICheck {
@@ -113,5 +113,9 @@ class UserGroupMembership implements ICheck {
 	public function isAvailableForScope(int $scope): bool {
 		// admin only by default
 		return $scope === IManager::SCOPE_ADMIN;
+	}
+
+	public function setEntitySubject(IEntity $entity, $subject): void {
+		// NOOP
 	}
 }
